+++ ipfabric/technology/managed_networks.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 
 from ipfabric import models
 
-logger = logging.getLogger()
+logger = logging.getLogger("python-ipfabric")
 
 
 class ManagedNetworks(BaseModel):