+++ ipfabric/tools/restore_intents.py
@@ -8,16 +8,36 @@ from typing import Any, Union
 
 from pydantic.dataclasses import dataclass
 
-logger = logging.getLogger()
+logger = logging.getLogger("python-ipfabric")
+
+INTENTS = "intents.json"
+GROUPS = "groups.json"
+DASHBOARD = "dashboard.json"
 
 V4_4_PATH = "ipfabric.tools.factory_defaults.v4.4"
+V5_0_PATH = "ipfabric.tools.factory_defaults.v5.0"
+V6_0_PATH = "ipfabric.tools.factory_defaults.v6.0"
 
 DEFAULT_SETUP = {
     "v4.4": {
-        "intents": json.loads(importlib.resources.read_text(V4_4_PATH, "intents.json")),
-        "groups": json.loads(importlib.resources.read_text(V4_4_PATH, "groups.json")),
+        "intents": json.loads(importlib.resources.read_text(V4_4_PATH, INTENTS)),
+        "groups": json.loads(importlib.resources.read_text(V4_4_PATH, GROUPS)),
         "dashboard": json.loads(
-            importlib.resources.read_text(V4_4_PATH, "dashboard.json"), object_pairs_hook=OrderedDict
+            importlib.resources.read_text(V4_4_PATH, DASHBOARD), object_pairs_hook=OrderedDict
+        ),
+    },
+    "v5.0": {
+        "intents": json.loads(importlib.resources.read_text(V5_0_PATH, INTENTS)),
+        "groups": json.loads(importlib.resources.read_text(V5_0_PATH, GROUPS)),
+        "dashboard": json.loads(
+            importlib.resources.read_text(V5_0_PATH, DASHBOARD), object_pairs_hook=OrderedDict
+        ),
+    },
+    "v6.0": {
+        "intents": json.loads(importlib.resources.read_text(V6_0_PATH, INTENTS)),
+        "groups": json.loads(importlib.resources.read_text(V6_0_PATH, GROUPS)),
+        "dashboard": json.loads(
+            importlib.resources.read_text(V6_0_PATH, DASHBOARD), object_pairs_hook=OrderedDict
         ),
     }
 }
@@ -52,7 +72,8 @@ class RestoreIntents:
             exit()
         self._restore(DEFAULT_SETUP[v]["intents"], DEFAULT_SETUP[v]["groups"], DEFAULT_SETUP[v]["dashboard"])
 
-    def _save_to_file(self, intents, groups, dashboard):
+    @staticmethod
+    def _save_to_file(intents, groups, dashboard):
         tstamp = datetime.now().strftime("%Y%m%d-%H%M%S")
 
         def save(name, data):